forecast plot is moved to graph_plot.py
